add: code refactored with MSC architecture and middleware layer added
