+++ main.go
@@ -1,8 +1,8 @@
 package main
 
 import (
-	"challenge/exchangeHandler"
-	"challenge/execSQL"
+	"challenge/controller"
+	"challenge/model"
 	"fmt"
 	"log"
 	"net/http"
@@ -15,9 +15,9 @@ func main() {
 
 	fs := http.FileServer(http.Dir("public"))
 
-	router.HandleFunc("/exchange/{amount}/{from}/{to}/{rate}", exchangeHandler.CurrencyConverter).Methods(http.MethodGet)
+	router.HandleFunc("/exchange/{amount}/{from}/{to}/{rate}", controller.CurrencyConverter).Methods(http.MethodGet)
+	router.HandleFunc("/exchange", model.GetAllConversions).Methods(http.MethodGet)
 	router.Handle("/", fs).Methods(http.MethodGet)
-	router.HandleFunc("/exchange", execSQL.GetAllConversions).Methods(http.MethodGet)
 
 	fmt.Println("Executando na porta 8000")
 	log.Fatal(http.ListenAndServe(":8000", router))
